requirements.txt
asgiref==3.3.1
appdirs==1.4.4
certifi==2019.11.28
cffi==1.14.5
chardet==3.0.4
cryptography==3.4.6
defusedxml==0.7.1
Deprecated==1.2.7
django-allauth==0.44.0
django-crispy-forms==1.11.0
django-environ==0.4.5
dj-database-url==0.5.0
Django==3.1.7
distlib==0.3.1
django-heroku==0.3.1
filelock==3.0.12
gitdb2==2.0.6
GitPython==3.0.5
gunicorn==20.0.4
idna==2.8
isodate==0.6.0
oauthlib==3.1.0
pipenv==2020.11.15
pycparser==2.20
pyparsing==2.4.7
PyGithub==1.45
PyJWT==1.7.1
python3-openid==3.2.0
psycopg2==2.8.6
psycopg2-binary==2.8.6
pytz==2021.1
requests==2.25.1
requests-oauthlib==1.3.0
rdflib==5.0.0
six==1.14.0
sqlparse==0.4.1
smmap2==2.0.5
toml==0.10.0
urllib3==1.25.7
virtualenv==20.4.3
virtualenv-clone==0.5.4
wrapt==1.11.2
whitenoise==5.2.0